Cleanup and reorganization
Creating a new process group in a pidns child doesn't seem to
work, so we create it in the parent. This may have the side
effect of SIGKILL rather than SIGHUP being sent to processes
in a pid ns, but it doesn't seem to work otherwise.

We also try to avoid forking twice.